+++ modules/compatibility/src/test/resources/docker/test-config.xml
